requirements.txt
amqp==5.0.5
asgiref==3.3.1
attrs==20.3.0
awsebcli==3.19.4
bcrypt==3.2.0
billiard==3.6.3.0
blessed==1.18.0
botocore==1.20.49
cached-property==1.5.2
celery==5.0.5
cement==2.8.2
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
colorama==0.4.3
coverage==5.5
cryptography==3.4.6
defusedxml==0.7.0rc2
Django==3.1.8
django-cors-headers==3.7.0
django-filter==2.4.0
django-rest-auth==0.9.5
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
docker==4.4.4
docker-compose==1.25.5
dockerpty==0.4.1
docopt==0.6.2
future==0.16.0
gunicorn==20.0.4
idna==2.10
jmespath==0.10.0
jsonschema==3.2.0
kombu==5.0.2
Markdown==3.3.4
oauthlib==3.1.0
paramiko==2.7.2
pathspec==0.5.9
Pillow==8.1.0
prompt-toolkit==3.0.17
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.1
PyNaCl==1.4.0
pyrsistent==0.17.3
python-dateutil==2.8.1
python3-openid==3.2.0
pytz==2021.1
PyYAML==5.3.1
redis==3.5.3
requests==2.25.1
requests-oauthlib==1.3.0
semantic-version==2.8.5
six==1.15.0
social-auth-app-django==4.0.0
social-auth-core==4.0.3
sorl-thumbnail==12.7.0
sqlparse==0.4.1
stripe==2.56.0
termcolor==1.1.0
texttable==1.6.3
urllib3==1.26.3
vine==5.0.0
wcwidth==0.2.5
websocket-client==0.58.0
